capstone/requirements.txt
tensorflow==2.4.0
streamlit==0.74.1
numpy==1.19.3
pandas==1.1.5
Pillow==8.1.0
